timeswitch: split time and state into different items
timeswitch now uses seperate items to set the time and state for items,
alternatively the "state" property can be used to set the state with not state item specified.
Note that if "state" is set (non empty string) it will override any state item specified.

The "disp/time" template was modified in accordance to now only show the time, not the state.